python_backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy

opencv-python-headless==4.8.1.78

torch==2.1.2
torchvision==0.16.2

ultralytics
